feat : awstrans.py 구조 변경
